requirements.txt
streamlit
langchain
langchain-community
chromadb
pypdf
openai
tiktoken
python-dotenv
sentence-transformers
